more cleanups + file split a little bit
