+++ comp/netflow/common/flow.go
@@ -131,8 +131,10 @@ var (
 	}
 )
 
-// AggregationHash return a hash used as aggregation key
-func (f *Flow) AggregationHash() uint64 {
+// PerReporterHash returns a hash that uniquely identifies a flow from a specific
+// reporter. It includes exporter address, namespace, and ingress interface so that
+// the same logical flow observed by different devices produces distinct keys.
+func (f *Flow) PerReporterHash() uint64 {
 	h := fnv.New64()
 	h.Write([]byte(f.Namespace))                           //nolint:errcheck
 	h.Write(f.ExporterAddr)                                //nolint:errcheck
@@ -146,12 +148,12 @@ func (f *Flow) AggregationHash() uint64 {
 	return h.Sum64()
 }
 
-// FlowKeyHash returns a hash of the flow's 5-tuple (src/dst addr, src/dst port,
-// IP protocol) plus ToS, used to group flows from multiple reporters together in
-// deduplication mode. Unlike AggregationHash, this excludes the exporter address,
-// namespace, and ingress interface so that the same flow observed by different
-// devices maps to the same key.
-func (f *Flow) FlowKeyHash() uint64 {
+// DeduplicationHash returns a hash of the flow's 5-tuple (src/dst addr, src/dst
+// port, IP protocol) plus ToS. Used to group flows from multiple reporters
+// together in deduplication mode. Unlike PerReporterHash, this excludes the
+// exporter address, namespace, and ingress interface so that the same flow
+// observed by different devices maps to the same key.
+func (f *Flow) DeduplicationHash() uint64 {
 	h := fnv.New64()
 	h.Write(f.SrcAddr)                                 //nolint:errcheck
 	h.Write(f.DstAddr)                                 //nolint:errcheck
@@ -162,9 +164,9 @@ func (f *Flow) FlowKeyHash() uint64 {
 	return h.Sum64()
 }
 
-// IsEqualFlowContext check if the flow and another flow have equal values for all fields used in `AggregationHash`.
-// This method is used for hash collision detection.
-func IsEqualFlowContext(a Flow, b Flow) bool {
+// IsEqualPerReporterContext checks if two flows have equal values for all fields
+// used in PerReporterHash. Used for hash collision detection.
+func IsEqualPerReporterContext(a Flow, b Flow) bool {
 	if a.Namespace == b.Namespace &&
 		bytes.Equal(a.ExporterAddr, b.ExporterAddr) &&
 		bytes.Equal(a.SrcAddr, b.SrcAddr) &&
